pyproject.toml
[tool.poetry]
name = "nonebot-plugin-who-at-me"
version = "0.1.0"
description = "Find who on earth has ated you"
authors = ["SEAFHMC <soku_ritsuki@outlook.com>"]
license = "MIT License"
readme = "README.md"
homepage = "https://github.com/SEAFHMC/nonebot-plugin-who-at-me"
repository = "https://github.com/SEAFHMC/nonebot-plugin-who-at-me"
documentation = "https://github.com/SEAFHMC/nonebot-plugin-who-at-me/blob/main/README.md"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = ">=2.0.0-beta.1, <=2.0.0-beta.3"
nonebot-adapter-onebot = ">=2.0.0-beta.1, <=2.1.0"
peewee = ">=3.14.4"
Pillow = ">=8.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
